third_party/blink/web_tests/external/wpt/html/syntax/charset/xhr-expected.txt

This is a testharness.js-based test.
[FAIL] Check after-head-after-1kb-crlf.html
  assert_equals: Check characterSet expected "UTF-8" but got "windows-1251"
Harness: the test ran to completion.
